Update tranche labels for new tranches
Some minor changes to the tranche labels:
- Add tranche 15 & 16, which were just created
- Remove label for tranche 10

Tranche 10’s label was true, but also contained other tokens that were issued unlocked, so it was deemed to be clearer to have it unlabelled than a semi-correct label, or another potentially misleading label. Checked this with @barnabee & @davidsiska-vega as they were in the room.